requirements.txt
Flask_Cors==3.0.3
Flask==0.12.2
gunicorn==19.9.0
python-dotenv==0.8.2
requests==2.18.4
psycopg2==2.7.4
psycopg2-binary==2.7.4
